Switch statement using JavaScript
first we put switch statement into the function named 'player'. Then give the test cases. The next function 'logger' where the output of the player function is returned and stored into a variable named 'played' and a message is passed with the output of the first function. At last the 'logger' function is called with an input based on given test cases.